refactor(tests): Extract TestHttpClient and TestResponse classes for better test structure
- Moved TestHttpClient and TestResponse classes to separate files for improved organization.
- Simplified NavigationServiceTest and PageServiceTest by removing inline class definitions.
- Enhanced test readability and maintainability by structuring test-specific clients separately.
