+++ tests/Service/PageServiceTest.php
@@ -4,112 +4,28 @@ declare(strict_types=1);
 
 namespace Dotcms\PhpSdk\Tests\Service;
 
-use Dotcms\PhpSdk\Config\Config;
 use Dotcms\PhpSdk\Exception\ResponseException;
-use Dotcms\PhpSdk\Http\HttpClient;
-use Dotcms\PhpSdk\Http\Response as DotcmsResponse;
 use Dotcms\PhpSdk\Model\PageAsset;
 use Dotcms\PhpSdk\Request\PageRequest;
 use Dotcms\PhpSdk\Service\PageService;
-use GuzzleHttp\Client;
 use GuzzleHttp\Handler\MockHandler;
-use GuzzleHttp\HandlerStack;
 use GuzzleHttp\Promise\PromiseInterface;
 use GuzzleHttp\Psr7\Response as GuzzleResponse;
 use PHPUnit\Framework\TestCase;
-use Psr\Http\Message\ResponseInterface;
-
-/**
- * Test-specific HttpClient that allows injecting a mock client
- */
-class TestHttpClient extends HttpClient
-{
-    private MockHandler $mockHandler;
-
-    public function __construct(MockHandler $mockHandler)
-    {
-        $config = new Config(
-            'https://demo.dotcms.com/api/v1',
-            'test-api-key',
-            [
-                'headers' => ['Content-Type' => 'application/json'],
-                'verify' => false,
-                'timeout' => 30,
-                'connect_timeout' => 5,
-                'http_errors' => false,
-            ]
-        );
-        $this->mockHandler = $mockHandler;
-        parent::__construct($config);
-    }
-
-    protected function createClient(): Client
-    {
-        $handlerStack = HandlerStack::create($this->mockHandler);
-
-        return new Client(['handler' => $handlerStack]);
-    }
-
-    public function request(string $method, string $uri, array $options = []): DotcmsResponse
-    {
-        $client = $this->createClient();
-        $response = $client->request($method, $uri, $options);
-
-        return new TestResponse($response);
-    }
-
-    public function requestAsync(string $method, string $uri, array $options = []): PromiseInterface
-    {
-        $client = $this->createClient();
-        $promise = $client->requestAsync($method, $uri, $options);
-
-        return $promise->then(function ($response) {
-            return new TestResponse($response);
-        });
-    }
-}
-
-/**
- * Custom Response class for testing
- */
-class TestResponse extends DotcmsResponse
-{
-    private ResponseInterface $originalResponse;
-
-    public function __construct(ResponseInterface $response)
-    {
-        $this->originalResponse = $response;
-        parent::__construct($response);
-    }
-
-    public function toArray(): array
-    {
-        $body = (string)$this->originalResponse->getBody();
-        $data = json_decode($body, true);
-
-        if (json_last_error() !== JSON_ERROR_NONE) {
-            throw new ResponseException('Invalid JSON response: ' . json_last_error_msg());
-        }
-
-        return $data;
-    }
-}
 
 class PageServiceTest extends TestCase
 {
-    private TestHttpClient $httpClient;
-
     private MockHandler $mockHandler;
+
+    private TestHttpClient $httpClient;
 
     private PageService $pageService;
 
     protected function setUp(): void
     {
         $this->mockHandler = new MockHandler();
-        $handlerStack = HandlerStack::create($this->mockHandler);
-        $client = new Client(['handler' => $handlerStack]);
 
-        // Create our test-specific HttpClient with the mock client
+        // Create our test-specific HttpClient with the mock handler
         $this->httpClient = new TestHttpClient($this->mockHandler);
         $this->pageService = new PageService($this->httpClient);
     }
